requirements.txt
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.4.0
pandas>=1.3.0
pyyaml>=5.4.0
pytest>=6.2.0
pytest-cov>=2.12.0
pytest-benchmark>=3.4.0
pytest-timeout>=2.1.0
pytest-xdist>=2.4.0
pytest-mock>=3.6.0
pytest-asyncio>=0.15.0
pytest-randomly>=3.8.0
pytest-repeat>=0.9.0
pytest-rerunfailures>=10.2
pytest-sugar>=0.9.4
pytest-html>=3.1.1
pytest-metadata>=1.11.0
pytest-profiling>=1.7.0
pytest-clarity>=1.0.1
pytest-verbose-parametrize>=1.7.0
pytest-env>=0.8.0
pytest-custom-exit-code>=0.3.0
pytest-instafail>=0.4.2
pytest-reportlog>=0.3.0
pytest-reportportal>=5.0.0
dash>=2.0.0
plotly>=5.3.0 